requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
email-validator==2.1.0
markupsafe==2.1.3
Jinja2==3.1.2
Werkzeug==2.3.7
click==8.1.7
blinker==1.7.0
itsdangerous==2.1.2python